- Removing not needed tasks. - General renames - Adding change file - Updating dependencies
